Remove red flag about user testing
While this red flag makes for a nice, pithy statement, it misunderstands linguistics. Breathalyzer tests, drug tests, etc. doesn't test the breathalyzer or the drug. Compounds like "_____ testing" name either the quality being tested (security, **usability**), the testing instrument (breathalyzer, **user**), the activity being tested (driving, crash), the target being tested, etc.

"User testing" occurs frequently in textbook, literature, articles, etc. – all of which understand perfectly that it is not the user being tested, but the application. They presumably are also sensitive to the flexible construction of compound nouns, and therefore find no issue with "user testing". This list item seems to just be a particular axe to grind, and is not a basis for a red flag.

If a submission really does misunderstand the purpose of usability testing, it will be evident and can be penalized for it – rather than penalizing a submission for a word choice that is valid.